GUI is practically complete. Only thing that need to be added is time base control for both osciloscope channels, but that is a minor change, probably add tomorrow.

+++ OsciGUI/OsciGUI_IdeaProject/src/sample/TimeControlSettings.java
@@ -1,5 +1,5 @@
 package sample;
 
 public class TimeControlSettings {
-    public double timePerDivision;
+    public float timePerDivision;
 }
